Merge From feature/watch/cssV6 into watch

See merge request s09-final/S09P31B208!286

+++ watch/app/src/main/java/com/dallim/service/TimerService.java
@@ -196,7 +196,7 @@ public class TimerService extends Service {
                 } else if (tempDistance >= 1000  && tempDistance >= lastTtsDistance + 1000){
                     lastTtsDistance = tempDistance;
                     tempDistance = Math.round(tempDistance/1000);
-                    ttsUtil.speak("현재 달린 거리는 " + String.valueOf(tempDistance)+ "km 입니다.");
+                    ttsUtil.speak("현재 달린 거리는 " + String.valueOf((int) tempDistance)+ "km 입니다.");
                 }
             }
         }
